+++ includes/Games/Tower_Defense/BaseTower.h
@@ -49,7 +49,7 @@ public:
 
 	double get_cooldown() { return m_cooldown / 1000.0; }
 	double get_distance() { return m_distance; }
-	unsigned int get_damage() { return m_damage; }
+	double get_damage() { return m_damage; }
 	unsigned int get_coast() { return m_coast; }
 	unsigned int get_upgrade_coast() { return m_coast_upgrade; }
 	virtual std::string get_custom() { return ""; };
@@ -79,7 +79,7 @@ protected:
 	double m_cooldown;
 	double m_distance;
 
-	unsigned int m_damage;
+	double m_damage;
 	unsigned int m_coast;
 	unsigned int m_coast_upgrade;
